rendering products in cart by them id

+++ src/components/Cart.tsx
@@ -2,6 +2,7 @@
 import { useShoppingCartStore } from '@/store/shoppingCart';
 import { ProductCard } from './ProductCard';
 import { ProductInCart } from './ProductInCart';
+import { products } from '@/mocks/products.json';
 
 export function Cart() {
   const cart = useShoppingCartStore((state) => state.cart);
@@ -9,16 +10,21 @@ export function Cart() {
 
   return (
     <section className='grid gap-2 lg:grid-cols-3 xl:grid-cols-4 flex-1 mt-5'>
-      {cart.map((item) => (
-        <ProductInCart
-          key={item.id}
-          handleDeleteItemAction={deleteItem}
-          itemId={item.id}
-          qty={item.qty}
-        >
-          <ProductCard {...item} />
-        </ProductInCart>
-      ))}
+      {cart.map((item) => {
+        const product = products.find((product) => product.id === item.id);
+        if (!product) return null;
+
+        return (
+          <ProductInCart
+            key={item.id}
+            handleDeleteItemAction={deleteItem}
+            itemId={item.id}
+            qty={item.qty}
+          >
+            <ProductCard {...product} />
+          </ProductInCart>
+        );
+      })}
     </section>
   );
 }
